Move test side-by-side to rule
